requirements.txt
appdirs==1.4.3
asn1crypto==1.0.1
astroid==2.3.1
attrs==19.2.0
black==19.3b0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.4
coveralls==1.8.2
cryptography==2.7
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
docopt==0.6.2
entrypoints==0.3
flake8==3.7.8
idna==2.8
isort==4.3.21
jwt==0.6.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.2
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
